Put route components in a separate file
